requirements.txt
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.5.2
asn1crypto==1.4.0
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.2
django-filter==22.1
djangorestframework==3.11.0
dnspython==2.1.0
docopt==0.6.2
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
Flask==2.2.2
future==0.18.2
gpg===1.16.0-unknown
gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
kazam==1.4.5
keyring==23.5.0
keyrings.alt==4.1.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauth==1.0.1
oauthlib==3.2.0
olefile==0.46
packaging==21.3
PAM==0.4.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipreqs==0.4.11
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
SecretStorage==3.3.1
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
ssh-import-id==5.11
system-service==0.3
systemd-python==234
terminator==2.1.1
Twisted==22.1.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
zope.interface==5.4.0
Django==4.1.2
psycopg2>=2.8